cobalt/heap: limit all sizes and offsets to 32bit values
The allocator cannot handle more than 2Gb of storage area by design,
and the uapi/ bits already assume that sizes and offsets produced by
Cobalt heaps fit in a 32bit word, for 32/64 ABI neutrality.

Therefore we may turn all relevant long types to u32 in the
implementation, to clearly state the addressing limit of this
allocator.

+++ kernel/drivers/ipc/iddp.c
@@ -160,7 +160,7 @@ static void __iddp_free_mbuf(struct iddp_socket *sk,
 }
 
 static void __iddp_flush_pool(struct xnheap *heap,
-			      void *poolmem, u_long poolsz, void *cookie)
+			      void *poolmem, u32 poolsz, void *cookie)
 {
 	free_pages_exact(poolmem, poolsz);
 }
